feat: 全ファイルにカスタムルールを適用する
