+++ eslint-local-rules/index.js
@@ -1,0 +1,8 @@
+'use strict'
+
+module.exports = {
+  'sample-rule': require('./sample-rule.js'),
+  'no-hoge': require('./no-hoge.js'),
+  'no-anpan': require('./no-anpan.js'),
+  'taboo-vars': require('./taboo-vars.js'),
+}
