+++ eslint-local-rules/no-hoge.js
