+++ package.json
@@ -40,6 +40,7 @@
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-filenames": "^1.3.2",
     "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-local-rules": "^2.0.0",
     "eslint-plugin-react-hooks": "^4.5.0",
     "eslint-plugin-unused-imports": "^2.0.0",
     "git-cz": "^4.9.0",
